% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_date.R, R/clean_colnames.R
\name{clean_colnames}
\alias{clean_colnames}
\title{Convert a timestamp to a date in a timezone defined}
\usage{
clean_colnames(df, ws_char = NA, to_lower = TRUE)

clean_colnames(df, ws_char = NA, to_lower = TRUE)
}
\arguments{
\item{df}{The dataframe for the columns to be modified in-place}

\item{ws_char}{character, length 1. The character to be set as the replacement for a space within the column name.}

\item{to_lower}{logical. Default is TRUE. If TRUE, the column names are set using tolower()}

\item{df}{The dataframe for the columns to be modified in-place}

\item{ws_char}{character, length 1. The character to be set as the replacement for a space within the column name.}

\item{to_lower}{logical. Default is TRUE. If TRUE, the column names are set using tolower()}
}
\description{
This function performs common

This function performs common
}
\examples{
clean_colnames()
clean_colnames()
}
