% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_project.R
\name{init_project}
\alias{init_project}
\title{Initialize a project structure within the current directory}
\usage{
init_project()
}
\description{
This function will initialize a project structure within the current
directory.  This includes files and directories for an arbitrary
data science project.  More than likely, not all are needed, but
remove as necessary.  If the directory or file exists, it will not
be modified.
}
\examples{
init_project()
}
