% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nhl_parse_pbp.R
\name{nhl_parse_pbp}
\alias{nhl_parse_pbp}
\title{Parse NHL play by play JSON files}
\usage{
nhl_parse_pbp(gid = "2019020010")
}
\arguments{
\item{gid}{character, length 1.  The NHL gameid}
}
\description{
This function parses the JSON file and returns a flattened dataframe for the plays and related data.
}
\examples{
nhl_parse_pbp()
}
