% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_rasa_nlu.R
\name{write_rasa_nlu}
\alias{write_rasa_nlu}
\title{Generate rasa NLU training data file}
\usage{
write_rasa_nlu(dat, path = "train.json")
}
\arguments{
\item{dat}{The dataframe holding the text data for questions}

\item{path}{character the path/name of the file to be exported. Default is in current directory with a name of train.json}
}
\description{
Take a dataframe and export a json dataset in rasa NLUs format.  The dataframe must contain a column
called text, which should represent the questions that you want to tag with intents and entities in the RASA NLU
training tool https://rasahq.github.io/rasa-nlu-trainer/.
https://nlu.rasa.com/dataformat.html
}
